Fix pipeline exposure launcher hang (#499)
When using pytorch, if you're going to play games with multiprocessing (in this case, loading a model in one process, then loading the same model in another process), you need to make sure pytorch is not using threading.  To my mind, this is very irritating behavior on the part of pytorch, but what can you do.

torch.set_num_threads( 1 )